+++ listenbrainz/webserver/static/js/src/RecentListens.tsx
@@ -21,7 +21,7 @@ import GlobalAppContext, { GlobalAppContextT } from "./GlobalAppContext";
 import {
   WithAlertNotificationsInjectedProps,
   withAlertNotifications,
-} from "./AlertNotificationsHOC";
+} from "./notifications/AlertNotificationsHOC";
 
 import APIServiceClass from "./APIService";
 import BrainzPlayer from "./brainzplayer/BrainzPlayer";
