+++ listenbrainz/webserver/static/js/src/huesound/ColorPlay.tsx
@@ -12,7 +12,7 @@ import GlobalAppContext, { GlobalAppContextT } from "../GlobalAppContext";
 import {
   WithAlertNotificationsInjectedProps,
   withAlertNotifications,
-} from "../AlertNotificationsHOC";
+} from "../notifications/AlertNotificationsHOC";
 
 import APIServiceClass from "../APIService";
 import BrainzPlayer from "../brainzplayer/BrainzPlayer";
