+++ listenbrainz/webserver/static/js/src/playlists/Playlist.tsx
@@ -26,7 +26,7 @@ import { Integrations } from "@sentry/tracing";
 import {
   withAlertNotifications,
   WithAlertNotificationsInjectedProps,
-} from "../AlertNotificationsHOC";
+} from "../notifications/AlertNotificationsHOC";
 import APIServiceClass from "../APIService";
 import GlobalAppContext, { GlobalAppContextT } from "../GlobalAppContext";
 import SpotifyAPIService from "../SpotifyAPIService";
