+++ listenbrainz/webserver/static/js/tests/notifications/AlertNotificationsHOC.test.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { shallow, mount } from "enzyme";
-import { withAlertNotifications } from "./AlertNotificationsHOC";
+import { withAlertNotifications } from "../../src/notifications/AlertNotificationsHOC";
 
 const fakeComponent = function FakeComponent() {
   return <div />;
